Source code/requirements.txt
absl-py==0.7.1
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
bz2file==0.98
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
dopamine-rl==3.0.1
entrypoints==0.3
Flask==1.1.2
future==0.17.1
gast==0.2.2
gensim==3.8.1
gevent==1.4.0
gin-config==0.3.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.25.0
gunicorn==20.0.4
gym==0.17.1
h5py==2.10.0
httplib2==0.18.1
idna==2.9
imbalanced-learn==0.6.2
imblearn==0.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mesh-tensorflow==0.1.16
mistune==0.8.4
mol2vec==0.1
mordred==1.2.0
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.5
notebook==6.0.3
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==3.4.4.19
opt-einsum==2.3.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.7
pypng==0.0.20
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.5.2
regex==2019.11.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
smart-open==1.8.4
statsmodels==0.11.1
sympy==1.6
tensor2tensor==1.15.7
tensorboard==2.1.0
tensorflow-addons==0.9.1
tensorflow-datasets==3.1.0
tensorflow-estimator==2.1.0
tensorflow-gan==2.0.0
tensorflow-gpu==2.1.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.2
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-slim==1.1.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typeguard==2.9.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
